refactor: 홈탭 코드 리팩토링
